sweeping changes, starting to refactor [wip]

+++ src/main/java/com/insidious/plugin/pojo/dao/TestCandidateMetadata.java
@@ -121,7 +121,7 @@ public class TestCandidateMetadata {
         if (testSubject == null) {
             return;
         }
-        this.testSubject_id = (long) testSubject.getValue();
+        this.testSubject_id = testSubject.getValue();
     }
 
     public long getCallTimeNanoSecond() {
